Added better naming and data typing system
Now the user will name the capture in the pattern and in the line FieldNames, the user will assign data types to the name of each capture.  Captures without names are not sent to InfluxDB, this allows for a more flexible pattern making as it makes grouping part of the pattern together without capture possible.